Extract text rendering logic into textRenderer utility
